+++ script/DeployCMTATWithWhitelist.s.sol
@@ -25,7 +25,7 @@ contract DeployCMTATWithWhitelist is Script {
         ICMTATConstructor.Engine memory engines = ICMTATConstructor.Engine(IRuleEngine(address(0)));
 
         token = new CMTATStandalone(forwarder, address(this), erc20Attributes, extraInformationAttributes, engines);
-        rule = new RuleWhitelist(admin, address(0), checkSpender);
+        rule = new RuleWhitelist(admin, address(0), checkSpender, false);
 
         token.setRuleEngine(IRuleEngine(address(rule)));