+++ test/RuleWhitelist/AccessControl/RuleWhitelistAccessControlOZ.t.sol
@@ -13,7 +13,7 @@ contract RuleWhitelistAccessControlOZ is Test, HelperContract, AccessControl {
     // Arrange
     function setUp() public {
         vm.prank(WHITELIST_OPERATOR_ADDRESS);
-        ruleWhitelist = new RuleWhitelist(WHITELIST_OPERATOR_ADDRESS, ZERO_ADDRESS, true);
+        ruleWhitelist = new RuleWhitelist(WHITELIST_OPERATOR_ADDRESS, ZERO_ADDRESS, true, false);
     }
 
     function testCanGrantRoleAsAdmin() public {