+++ test/RuleWhitelist/RuleWhitelistDeployment.t.sol
@@ -20,7 +20,7 @@ contract RuleWhitelistDeploymentTest is Test, HelperContract {
         MinimalForwarderMock forwarder = new MinimalForwarderMock();
         forwarder.initialize(ERC2771ForwarderDomain);
         vm.prank(WHITELIST_OPERATOR_ADDRESS);
-        ruleWhitelist = new RuleWhitelist(WHITELIST_OPERATOR_ADDRESS, address(forwarder), true);
+        ruleWhitelist = new RuleWhitelist(WHITELIST_OPERATOR_ADDRESS, address(forwarder), true, false);
 
         // assert
         resBool = ruleWhitelist.hasRole(ADDRESS_LIST_ADD_ROLE, WHITELIST_OPERATOR_ADDRESS);
@@ -38,6 +38,6 @@ contract RuleWhitelistDeploymentTest is Test, HelperContract {
         forwarder.initialize(ERC2771ForwarderDomain);
         vm.expectRevert(AccessControlModuleStandalone.AccessControlModuleStandalone_AddressZeroNotAllowed.selector);
         vm.prank(WHITELIST_OPERATOR_ADDRESS);
-        ruleWhitelist = new RuleWhitelist(address(0), address(forwarder), true);
+        ruleWhitelist = new RuleWhitelist(address(0), address(forwarder), true, false);
     }
 }